--- packages/app/component/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "@exam/component",
-  "type": "module"
-}
