+++ packages/app/components/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@exam/components",
+  "type": "module"
+}
